fix dismissing notifications when you enter the web convo

+++ api/src/main/java/xyz/klinker/messenger/api/service/AccountService.java
@@ -52,6 +52,6 @@ public interface AccountService {
     Object updateSetting(@Query("account_id") String accountId, @Query("pref") String pref, @Query("type") String type, @Query("value") Object value);
 
     @POST("accounts/dismissed_notification")
-    Object dismissedNotification(@Query("account_id") String accountId, @Query("id") int id);
+    Object dismissedNotification(@Query("account_id") String accountId, @Query("id") long id);
 
 }
